DrawCheckBox replaced by structure checkbox, huge code refactoring
git-svn-id: svn://kolibrios.org@7244 a494cfbc-eb01-0410-851d-a64ba20cac60

+++ programs/cmm/eolite/include/about.h
@@ -4,7 +4,7 @@
 
 void about_dialog()
 {   
-	byte id;
+	int id;
 	proc_info about_form;
 	int about_x;
