+++ programs/cmm/taskbar2/taskbar2.c
@@ -59,7 +59,7 @@ dword COLOR_TEXT    = 0xFFFfff;
 
 void main()
 {
-	byte btn;
+	int btn;
 	load_dll(libio,  #libio_init,1);
 	load_dll(libimg, #libimg_init,1);
 	load_dll(libini, #lib_init,1);
